fix compile bugs & reviews
