+++ csrc/CMakeLists.txt
@@ -9,7 +9,8 @@ FILE(GLOB OP_SRCS
     ${PROJECT_OP_SRC_BASE}/alloc_extend/op_host/alloc_extend_tiling.cpp
     ${PROJECT_OP_SRC_BASE}/assign_cache_op/op_host/assign_cache.cpp
     ${PROJECT_OP_SRC_BASE}/mla_preprocess/op_host/mla_preprocess.cpp
-    ${PROJECT_OP_SRC_BASE}/pp_matmul_einsum/op_host/tiling/tiling_data.cpp
+    ${PROJECT_OP_SRC_BASE}/batch_matmul_transpose/op_host/batch_matmul_transpose.cpp
+    ${PROJECT_OP_SRC_BASE}/batch_matmul_transpose/op_host/tiling/tiling_data.cpp
     )
 
 # set the so name
@@ -20,7 +21,7 @@ ascendc_library(no_workspace_kernel STATIC
     ${PROJECT_OP_SRC_BASE}/helloworld/op_kernel/kernel_helloworld.cpp
     ${PROJECT_OP_SRC_BASE}/cache_location_assign/op_kernel/cache_loc_assign_kernel.cpp
     ${PROJECT_OP_SRC_BASE}/assign_cache_op/op_kernel/assign_cache_op.cpp
-    ${PROJECT_OP_SRC_BASE}/pp_matmul_einsum/op_kernel/pp_matmul_einsum_kernel.cpp
+    ${PROJECT_OP_SRC_BASE}/batch_matmul_transpose/op_kernel/batch_matmul_transpose_kernel.cpp
 )
 
 ascendc_library(workspace_kernel STATIC
