+++ csrc/utils/common.h
@@ -12,6 +12,8 @@
 #ifndef UTILS_COMMON_H
 #define UTILS_COMMON_H
 
+namespace host_utils{
+
 constexpr uint32_t BLK_SIZE_ALIN_FOR_INT64 = 4;
 constexpr uint32_t BLK_SIZE_ALIN_FOR_INT32 = 8;
 
@@ -25,4 +27,31 @@ inline uint64_t alinInt32Count(uint64_t count)
     return (count + BLK_SIZE_ALIN_FOR_INT32 - 1) / BLK_SIZE_ALIN_FOR_INT32 * BLK_SIZE_ALIN_FOR_INT32;
 }
 
+template <typename T>
+inline T CeilDiv(const T dividend, const T divisor)
+{
+    if (divisor == 0) {
+        return UINT32_MAX;
+    }
+    return (dividend + divisor - 1) / divisor;
+}
+
+template <typename T>
+inline T RoundUp(const T val, const T align = 16)
+{
+    if (align == 0 || val + align - 1 < val) {
+        return 0;
+    }
+    return (val + align - 1) / align * align;
+}
+
+template <typename T>
+inline T RoundDown(const T val, const T align = 16)
+{
+    if (align == 0) {
+        return 0;
+    }
+    return val / align * align;
+}
+}
 #endif  // UTILS_COMMON_H
